Add Tasks.seqActions(...) to sequence actions
This new task makes it easy to sequence a set of actions when you don't
need the result from the last task. Our current Tasks.seq(...) tasks
return the value of the last task. When recursively sequencing fan out
(parallel tasks) this can lead to a List<List<...List<T>...>> type for
the seq task when subsequent aggregation does not occur.

The behavior of seqActions differs from seq in that any task that fails
in the sequence causes the sequence to terminate immediately with an
error.