Don't use `alert` for trace text errors
Our previous tracevis required the user to hit a button to trigger
re-rendering of a trace. The current tracevis will try to re-render as
the text changes, so using `alert` becomes quite disruptive. Instead
this change will make the text in the text area red, make the hover for
the text area indicate the error, and log the error to the console (via
throw).