Don't hold ref to scheduled task after it has been cancelled
We have observed that tasks scheduled with the ScheduledExecutorService
remain reachable from that executor even after they have been cancelled.
This change aggresively eliminates the path from the scheduled executor
service to the underlying task by using a reference in between that will
be nulled out when the task has either run or been cancelled.