Added RequestContext provider to ParSeqRestClient.
RequestContext provider is a function that takes Request<?> object and
returns RequestContext object. It is used to create RequestContext
object in case when caller did not provide one as a parameter.

This change is backwards compatible. Up until now the implicit
RequestContext provider was: () -> new RequestContext() and it is now
explicit.