Avoid blocking in ASMBasedTaskDescriptor (#336)
Currently the ASMBasedTaskDescriptor uses ByteBuddy to hook into
the class loader to analyze bytecode so that ParSeq can generate
descriptions of tasks that were created via lambda functions to
aid in debugging and tracing.

In production, blocking on the lambda can cause a global lockout
of threads. This can happen because of the common fork join pool
being starved or because the classes being analyzed take a while.

If a class has already been analyzed, we do not need to await
the latch because we have a thread safe collection to access.
Classes can be analyzed multiple times in a JVM lifecycle because
of class unloading, but we don't care about subsequent analyses,
(the bytecode should not change with load -> unload -> load).