Create defined class to perform analysis within thread (#338)
When refactoring from an Anonymous class to a lambda, an infinite
recursion occurred because each analysis was required to then
fork and create a new analysis when it was done because it defined
a new anonymous class.

This change defines the class specifically to avoid it analyzing
itself. A follow up should be to filter anonymous classes unrelated
to ParSeq.